pulled stress recon and metric evaluation into their own tests
