+++ src/pyvale/vfm/optimiser.py
@@ -41,7 +41,6 @@ def evaluate_candidate(
     objective_function: IObjectiveFunction,
     experiment_data: ExperimentData,
 ) -> float | npt.NDArray[np.float64]:
-    
     print("Evaluating candidate with degrees of freedom:", degrees_of_freedom)
     updated_spatial_parameterisations = unpack_spatial_parameterisations(
         spatial_parameterisations,
