+++ src/pyvale/vfm/spatialparamhomogeneous.py
@@ -9,9 +9,6 @@ from pyvale.vfm.dof import DegreeOfFreedom
 from pyvale.vfm.spatialparam import ISpatialParameterisation
 
 
-# TODO: How do I flag whether value should be a dof or fixed?
-#   Maybe a constructor with a is_value_fixed var which informs
-#   how we should construct
 @dataclass(slots=True)
 class HomogeneousSpatialParameterisation(ISpatialParameterisation):
     value: float | DegreeOfFreedom | None = None
